Adds 'let' keyword to lexer
Moves the 'dofile' abstraction up one level. Adds a new function GutDoFile
which takes a GutState reference and a filename and lexes the file.

Next step will be to start the parser, which will ask for tokens as it
goes. Also need to implement the peek function for the lexer so that if
needed the parser can ask for a peek.