Adds a whole lot of stuff
Adds a core initialization function. For now this is only meant to add all keywords to the global string table.
Later this might be used for more stuff, maybe a shortcut for starting up tables, arrays, and, strings?

Adds a Keyword struct with the length of each keyword. Just as an easy way of starting up the string table.

When the string table is started up, all strings should be hashed then compared to check if they are keywords.
Might have an option to turn this off or change out the table.