Adds lots of small changes
Changes the instruction format to be 32 bits wide.

Removes variables from functions because we only need to keep them around
for parsing. Will probably be added to the prototype struct in
gut_parser.h.

Adds a test.c file for starting work on the vm. Not sure if the best
approach is getting the parser to work before starting on the vm or simply
starting on the vm right away.

Removes most dependencies from guttural.h. It should be super easy to
include guttural into an existing project. Simply including guttural.h
should be all you need to do.